+++ montepy/input_parser/input_syntax_reader.py
@@ -146,6 +146,7 @@ def read_data(fh, mcnp_version, block_type=None, recursion=False):
 
     def flush_block():
         nonlocal block_counter, block_type
+        # keep parsing while there is input or termination has not been triggered
         if len(input_raw_lines) > 0:
             yield from flush_input()
         block_counter += 1
@@ -154,6 +155,17 @@ def read_data(fh, mcnp_version, block_type=None, recursion=False):
 
     def flush_input():
         nonlocal input_raw_lines
+        # IF 3  BLOCKS are parsed, the rest should be ignored with a warning and print 3 lines
+        if block_counter >= 3:
+            joined_lines = "\n".join(input_raw_lines[0:3])
+            msg = f"Unexpected input after line {current_file.lineno - 1}\n line content: {joined_lines}\n"
+            warnings.warn(
+                msg,
+                UndefinedBlock,
+                stacklevel=6,
+            )
+            return
+
         start_line = current_file.lineno + 1 - len(input_raw_lines)
         input = Input(
             input_raw_lines,
@@ -191,6 +203,7 @@ def read_data(fh, mcnp_version, block_type=None, recursion=False):
             and input_raw_lines
         ):
             yield from flush_input()
+
         # die if it is a vertical syntax format
         start_o_line = line[0:BLANK_SPACE_CONTINUE]
         # eliminate comments, and inputs that use # for other syntax
@@ -215,11 +228,13 @@ def read_data(fh, mcnp_version, block_type=None, recursion=False):
         old_line = line
         line = line[:line_length]
         if len(old_line) != len(line):
-            if len(line.split("$")[0]) >= line_length and not COMMENT_FINDER.match(
+            comment_free = old_line.split("$")[0]
+            if len(comment_free.rstrip()) > line_length and not COMMENT_FINDER.match(
                 line
             ):
                 warnings.warn(
-                    f"The line: {old_line} exceeded the allowed line length of: {line_length} for MCNP {mcnp_version}",
+                    f"The line number {fh.lineno} exceeded the allowed line length of: {line_length} for MCNP{mcnp_version} "
+                    f'and "{comment_free[line_length -1:].rstrip()}" was removed.',
                     LineOverRunWarning,
                 )
             # if extra length is a comment keep it long