+++ montepy/mcnp_object.py
@@ -4,7 +4,6 @@ from abc import ABC, ABCMeta, abstractmethod
 import copy
 import functools
 import itertools as it
-import sys
 import textwrap
 from typing import Union
 import warnings
@@ -188,15 +187,10 @@ class MCNP_Object(ABC, metaclass=_ExceptionContextAdder):
         if key.startswith("_"):
             super().__setattr__(key, value)
         else:
-            # kwargs added in 3.10
-            if sys.version_info >= (3, 10):
-                raise AttributeError(
-                    f"'{type(self).__name__}' object has no attribute '{key}'",
-                    obj=self,
-                    name=key,
-                )
             raise AttributeError(
                 f"'{type(self).__name__}' object has no attribute '{key}'",
+                obj=self,
+                name=key,
             )
 
     def _jit_light_init(self, input: Input):
@@ -530,6 +524,13 @@ The new input was:\n\n"""
 
     @_problem.setter
     def _problem(self, problem):
+        """
+        The problem this object is associated with if any.
+
+        Returns
+        -------
+        montepy.MCNP_Problem | None
+        """
         if problem is None:
             self._problem_ref = None
             return
@@ -551,6 +552,9 @@ The new input was:\n\n"""
         return self._tree.get_trailing_comment()
 
     def _delete_trailing_comment(self):
+        """
+        Deletes trailing comments from an object when it has been moved to another object.
+        """
         self._tree._delete_trailing_comment()
 
     def _grab_beginning_comment(self, padding: list[PaddingNode], last_obj=None):