+++ montepy/numbered_object_collection.py
@@ -20,6 +20,7 @@ def _enforce_positive(self, num):
 class NumberedObjectCollection(ABC):
     """A collections of MCNP objects.
 
+
     .. _collect ex:
 
     Examples
@@ -243,7 +244,6 @@ class NumberedObjectCollection(ABC):
         """
         if not isinstance(number, Integral):
             raise TypeError("The number must be an int")
-
         if number < 0:
             raise ValueError(f"The number must be non-negative. {number} given.")
         conflict = False
@@ -359,7 +359,9 @@ class NumberedObjectCollection(ABC):
             the object to delete
         """
         if not isinstance(delete, self._obj_class):
-            raise TypeError("")
+            raise TypeError(
+                f"Expected {self._obj_class.__name__}. {delete} of type: {type(delete).__name__} given."
+            )
         candidate = self[delete.number]
         if delete is candidate:
             del self[delete.number]
@@ -632,14 +634,16 @@ class NumberedObjectCollection(ABC):
         If starting_number, or step are not specified :func:`starting_number`,
         and :func:`step` are used as default values.
 
-
         .. versionchanged:: 0.5.0
             In 0.5.0 the default values were changed to reference :func:`starting_number` and :func:`step`.
+
+        .. versionchanged:: 1.2.0
+            start_num is now only a suggestion for the starting point. The returned number is not guaranteed to be start_num, but will be the next available number after start_num (or after the last assigned number).
 
         Parameters
         ----------
         start_num : int
-            the starting number to check.
+            Suggested starting number to check. Not guaranteed to be the returned value.
         step : int
             the increment to jump by to find new numbers.
 
@@ -656,13 +660,21 @@ class NumberedObjectCollection(ABC):
             start_num = self.starting_number
         if step is None:
             step = self.step
-        number = start_num
+        try:
+            self.check_number(start_num)
+            return start_num
+        except NumberConflictError:
+            pass
+        # Increment to next available number. If not set use start_num as is
+        last_assigned = getattr(self, "_last_assigned_number", start_num - step) + step
+        number = max(start_num, last_assigned)
         while True:
             try:
                 self.check_number(number)
                 break
             except NumberConflictError:
                 number += step
+        self._last_assigned_number = number
         return number
 
     def next_number(self, step=1):
@@ -1136,6 +1148,19 @@ class NumberedObjectCollection(ABC):
 
 
 class NumberedDataObjectCollection(NumberedObjectCollection):
+    """
+    This is a an abstract collection for numbered objects that are also Data Inputs.
+
+    This collection can be sliced to get a subset of the numberedDataObjectCollection.
+    Slicing is done based on the numberedDataObjectCollection numbers, not their order in the input.
+    For example, ``problem.numberedDataObjectCollection[1:3]`` will return a new `numberedDataObjectCollection`
+    containing numberedDataObjectCollection with numbers from 1 to 3, inclusive.
+
+    See also
+    --------
+    :class:`~montepy.numbered_object_collection.NumberedObjectCollection`
+    """
+
     def __init__(self, obj_class, objects=None, problem=None):
         self._last_index = None
         if problem and objects: